+++ app/src/main/java/com/codesoom/assignment/controllers/ControllerCommonErrorAdvice.java
@@ -1,0 +1,13 @@
+package com.codesoom.assignment.controllers;
+
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+/**
+ * 공통 예외 처리 핸들러
+ */
+@ResponseBody
+@ControllerAdvice
+public class ControllerCommonErrorAdvice {
+    // TODO: if common handling of exceptions required, fill in!
+}
