+++ app/src/main/java/com/codesoom/assignment/controllers/ControllerProductErrorAdvice.java
@@ -1,0 +1,22 @@
+package com.codesoom.assignment.controllers;
+
+import com.codesoom.assignment.dto.ErrorResponse;
+import com.codesoom.assignment.errors.ProductNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+/**
+ * 제품에 관한 예외 처리 핸들러
+ */
+@ResponseBody
+@ControllerAdvice
+public class ControllerProductErrorAdvice {
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(ProductNotFoundException.class)
+    public ErrorResponse handleProductNotFound() {
+        return new ErrorResponse("Product not found");
+    }
+}
